New Datasource: github_repositories_v2
The `github_repositories` datasource used the Github Search API, in
order to return a list of repositories names (or theirs full names).
The datasource, will return up to 1k repositories as governed by
the API limits.

Currently, the datasource doesn't support return more than 1k
repositories. Also, the list only returns the repositories names
in a list of string form. In order have access to the repository
attributes, one could use the pair the resource with the
`github_repository` to fetch each repository information. This
leads to lots of API calls for a organization with lots of
repositories.

Therefore, let's introduce a new data source, namely
`github_repository_v2`, that will list all repositories in
organization and return few exported attributes to that can be
used in the Github repositories. As the resource behaves very
differently from the `github_repository` a new name is chosen;
which will, also, ensure a good backward-compatible feature.